refactor: rename from .iterface to .types
