+++ packages/web/src/common/repositories/event/local.event.repository.ts
@@ -8,7 +8,7 @@ import {
 import { getOfflineDataStore } from "@web/common/storage/offline-data/offline-data.store.registry";
 import { preserveLocalEventMarker } from "@web/common/storage/types/local-event.types";
 import { type Response_GetEventsSuccess } from "@web/events/event.types";
-import { type EventRepository } from "./event.repository.interface";
+import { type EventRepository } from "./event.repository.types";
 
 /**
  * Local event repository implementation using the offline data store.
