+++ packages/web/src/common/repositories/event/remote.event.repository.ts
@@ -10,7 +10,7 @@ import {
 } from "@web/common/apis/util/backend-unavailable-error.util";
 import { EventApi } from "@web/events/event.api";
 import { type Response_GetEventsSuccess } from "@web/events/event.types";
-import { type EventRepository } from "./event.repository.interface";
+import { type EventRepository } from "./event.repository.types";
 import { LocalEventRepository } from "./local.event.repository";
 
 export class RemoteEventRepository implements EventRepository {
