+++ packages/web/src/events/mutations/useEventMutations.ts
@@ -6,8 +6,8 @@ import {
   type Schema_Event,
 } from "@core/types/event.types";
 import { type EventRepositorySource } from "@web/common/repositories/event/event.repository.factory";
-import { type EventRepository } from "@web/common/repositories/event/event.repository.interface";
 import { useEventRepositorySource } from "@web/common/repositories/event/event.repository.source.store";
+import { type EventRepository } from "@web/common/repositories/event/event.repository.types";
 import { getEventRepositoryBySource } from "@web/common/repositories/event/event.repository.util";
 import { handleError } from "@web/common/utils/event/event.util";
 import { createObjectIdString } from "@web/common/utils/id/object-id.util";
