+++ packages/web/src/events/queries/week.event.query.ts
@@ -1,4 +1,4 @@
-import { type EventRepository } from "@web/common/repositories/event/event.repository.interface";
+import { type EventRepository } from "@web/common/repositories/event/event.repository.types";
 import { type Payload_GetEvents } from "@web/events/event.types";
 import { EventDateUtils, normalizeEventList } from "./event.query.normalize";
 import { type NormalizedEventQueryData } from "./event.query.types";
